# .sdd/contracts/api/web-api.openapi.yaml
openapi: 3.1.0
info:
  title: "SDD Web API"
  version: "0.3.0"
  description: |
    REST-API des SDD Web UI Servers (SPEC-0003, CON-0007).
    Läuft lokal auf http://localhost:8000. Kein Auth erforderlich.
  contact:
    name: "Boris"
servers:
  - url: "http://localhost:8000/api"
    description: "Lokaler Dev-Server"
paths:
...
  /contracts:
    get:
      operationId: listContracts
      summary: "Alle Contracts auflisten"
      responses:
        "200":
          description: "Liste aller Contracts"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"

    post:
      operationId: createContract
      summary: "Neuen Contract anlegen"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractCreate"
      responses:
        "201":
          description: "Contract angelegt"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedArtifact"
  /contracts/{contract_id}:
    get:
      operationId: getContract
      summary: "Contract mit Body und Artifact-Inhalt abrufen"
      parameters:
        - $ref: "#/components/parameters/ContractId"
      responses:
        "200":
          description: "Contract inkl. Body"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Tests ─────────────────────────────────────────────────────────────────
  /tests:
    get:
      operationId: listTests
      summary: "Alle Tests auflisten"
      responses:
        "200":
          description: "Liste aller Tests"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"

    post:
      operationId: createTest
      summary: "Neuen Test anlegen"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestCreate"
      responses:
        "201":
          description: "Test angelegt"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedArtifact"
...
  /open:
    post:
      operationId: openInEditor
      summary: "Datei in VS Code öffnen"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenRequest"
      responses:
        "200":
          description: "Datei geöffnet"
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:     { type: boolean }
                  opened: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /ai/generate-spec:
    post:
      operationId: aiGenerateSpec
      summary: "Spec aus Titel + Kontext generieren"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiGenerateSpec"
      responses:
        "200":
          description: "Generierter Spec-Text"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponse"
  /ai/improve-spec:
    post:
      operationId: aiImproveSpec
      summary: "Bestehende Spec nach Anweisung verbessern"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiImproveSpec"
      responses:
        "200":
          description: "Verbesserter Spec-Text"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponse"
  /ai/suggest-contracts:
    post:
      operationId: aiSuggestContracts
      summary: "Contract-Typen für eine Spec vorschlagen"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSuggestContracts"
      responses:
        "200":
          description: "Vorgeschlagene Contracts"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponse"
...
components:
  parameters:
...
    ContractId:
      name: contract_id
      in: path
      required: true
      schema: { type: string, pattern: "^CON-[0-9]{4,}$" }
...
  schemas:
...
    Contract:
      type: object
      required: [id, title, type, format, spec, status]
      properties:
        id:           { type: string, example: "CON-0001" }
        title:        { type: string }
        type:         { type: string, enum: [api, behavior, data, performance] }
        format:       { type: string }
        spec:         { type: string }
        status:       { type: string, enum: [draft, active, deprecated] }
        version:      { type: string }
        artifact:     { type: string }
        abs_artifact: { type: string }
        tests:        { type: array, items: { type: string } }
        file:         { type: string }
        abs_file:     { type: string }
    ContractDetail:
      allOf:
        - $ref: "#/components/schemas/Contract"
        - type: object
          properties:
            body:             { type: string }
            artifact_content: { type: string, nullable: true }
    ContractCreate:
      type: object
      required: [spec_id, format]
      properties:
        spec_id: { type: string }
        format:  { type: string }
        title:   { type: string, default: "" }
    Test:
      type: object
      required: [id, title, level, spec, contract, status]
      properties:
        id:        { type: string, example: "TST-0001" }
        title:     { type: string }
        level:     { type: string, enum: [unit, integration, contract, acceptance, performance, property] }
        spec:      { type: string }
        contract:  { type: string }
        status:    { type: string, enum: [planned, implemented, passing, failing, skipped] }
        framework: { type: string }
        file:      { type: string }
        abs_file:  { type: string }
...
    TestCreate:
      type: object
      required: [spec_id, contract_id, level]
      properties:
        spec_id:     { type: string }
        contract_id: { type: string }
        level:       { type: string, enum: [unit, integration, contract, acceptance, performance, property] }
        title:       { type: string, default: "" }
    CreatedArtifact:
      type: object
      required: [id, file]
      properties:
        id:       { type: string }
        file:     { type: string }
        abs_file: { type: string }
...
    OpenRequest:
      type: object
      required: [abs_file]
      properties:
        abs_file: { type: string }
        line:     { type: integer, default: 1 }
    AiGenerateSpec:
      type: object
      required: [title]
      properties:
        title:       { type: string }
        description: { type: string, default: "" }
        context:     { type: string, default: "" }
    AiImproveSpec:
      type: object
      required: [spec_id, current_content, instructions]
      properties:
        spec_id:         { type: string }
        current_content: { type: string }
        instructions:    { type: string }
    AiSuggestContracts:
      type: object
      required: [spec_id, spec_content]
      properties:
        spec_id:      { type: string }
        spec_content: { type: string }
    AiUsageEntry:
      type: object
      properties:
        ts:                    { type: string, format: date-time }
        provider:              { type: string }
        operation:             { type: string }
        model:                 { type: string }
        input_tokens:          { type: integer }
        output_tokens:         { type: integer }
        cache_creation_tokens: { type: integer }
        cache_read_tokens:     { type: integer }
        cost_usd:              { type: number }
    AiResponse:
      type: object
      required: [result, usage]
      properties:
        result: { type: string }
        usage:  { $ref: "#/components/schemas/AiUsageEntry" }
...
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
  responses:
    NotFound:
      description: "Nicht gefunden"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ServiceUnavailable:
      description: "Externer Dienst nicht verfügbar (z.B. VS Code nicht gefunden)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }